backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
pytesseract
Pillow
pdf2image
opencv-python
numpy
